Correct components
