+++ .eslintrc.json
@@ -14,7 +14,8 @@
     "no-shadow": "off",
     "no-param-reassign": "off",
     "eol-last": "off",
-    "no-unused-vars": "off"
+    "no-unused-vars": "off",
+    "arrow-body-style": "off"
   },
   "ignorePatterns": [
     "dist/",
