+++ src/components/Main.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Books from '../redux/books/Books';
-import AddBook from '../redux/books/AddBook';
+import Books from './Books';
+import AddBook from './AddBook';
 
 function Main() {
   return (
